Move options to its own file
Purely moving code
